requirements.txt
pyrate-limiter==2.9.0
requests==2.28.2
streamlit==1.17.0
sqlalchemy<2.0
apache-airflow==2.5.1
dbt-core==1.4.1
dbt-sqlite==1.4.0
pytest==7.2.2
gspread==5.7.2
python-dotenv==1.0.0
